+++ relay-config/src/config.rs
@@ -569,28 +569,11 @@ pub struct Metrics {
     pub default_tags: BTreeMap<String, String>,
     /// Tag name to report the hostname to for each metric. Defaults to not sending such a tag.
     pub hostname_tag: Option<String>,
-    /// Global sample rate for all emitted metrics between `0.0` and `1.0`.
-    ///
-    /// For example, a value of `0.3` means that only 30% of the emitted metrics will be sent.
-    /// Defaults to `1.0` (100%).
-    pub sample_rate: f64,
     /// Interval for periodic metrics emitted from Relay.
     ///
     /// Setting it to `0` seconds disables the periodic metrics.
     /// Defaults to 5 seconds.
     pub periodic_secs: u64,
-    /// Whether local metric aggregation using statdsproxy should be enabled.
-    ///
-    /// Defaults to `true`.
-    pub aggregate: bool,
-    /// Allows emission of metrics with high cardinality tags.
-    ///
-    /// High cardinality tags are dynamic values attached to metrics,
-    /// such as project IDs. When enabled, these tags will be included
-    /// in the emitted metrics. When disabled, the tags will be omitted.
-    ///
-    /// Defaults to `false`.
-    pub allow_high_cardinality_tags: bool,
 }
 
 impl Default for Metrics {
@@ -601,10 +584,7 @@ impl Default for Metrics {
             prefix: "sentry.relay".into(),
             default_tags: BTreeMap::new(),
             hostname_tag: None,
-            sample_rate: 1.0,
             periodic_secs: 5,
-            aggregate: true,
-            allow_high_cardinality_tags: false,
         }
     }
 }
@@ -1224,9 +1204,9 @@ pub struct Processing {
     ///
     /// Must be between `0.0` and `1.0`, by default there is no limit configured.
     pub quota_cache_max: Option<f32>,
-    /// Configuration for attachment uploads.
-    #[serde(default)]
-    pub upload: UploadServiceConfig,
+    /// Configuration for the objectstore service.
+    #[serde(default, alias = "upload")]
+    pub objectstore: ObjectstoreServiceConfig,
 }
 
 impl Default for Processing {
@@ -1247,7 +1227,7 @@ impl Default for Processing {
             max_rate_limit: default_max_rate_limit(),
             quota_cache_ratio: None,
             quota_cache_max: None,
-            upload: UploadServiceConfig::default(),
+            objectstore: ObjectstoreServiceConfig::default(),
         }
     }
 }
@@ -1296,10 +1276,10 @@ impl Default for OutcomeAggregatorConfig {
     }
 }
 
-/// Configuration values for attachment uploads.
+/// Configuration values for the objectstore service.
 #[derive(Serialize, Deserialize, Debug)]
 #[serde(default)]
-pub struct UploadServiceConfig {
+pub struct ObjectstoreServiceConfig {
     /// The base URL for the objectstore service.
     ///
     /// This defaults to [`None`], which means that the service will be disabled,
@@ -1313,7 +1293,7 @@ pub struct UploadServiceConfig {
     pub timeout: u64,
 }
 
-impl Default for UploadServiceConfig {
+impl Default for ObjectstoreServiceConfig {
     fn default() -> Self {
         Self {
             objectstore_url: None,
@@ -1643,6 +1623,28 @@ impl Default for Cogs {
     }
 }
 
+/// Configuration for the upload service.
+#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
+#[serde(default)]
+pub struct Upload {
+    /// Maximum number of uploads that the service accepts.
+    ///
+    /// Additional uploads will be rejected.
+    pub max_concurrent_requests: usize,
+    /// Maximum time spent trying to upload, in seconds.
+    /// Currently only used by non-processing relays, as the objectstore service has its own timeout.
+    pub timeout: u64,
+}
+
+impl Default for Upload {
+    fn default() -> Self {
+        Self {
+            max_concurrent_requests: 10,
+            timeout: 60,
+        }
+    }
+}
+
 #[derive(Serialize, Deserialize, Debug, Default)]
 struct ConfigValues {
     #[serde(default)]
@@ -1683,6 +1685,8 @@ struct ConfigValues {
     health: Health,
     #[serde(default)]
     cogs: Cogs,
+    #[serde(default)]
+    upload: Upload,
 }
 
 impl ConfigObject for ConfigValues {
@@ -2191,21 +2195,6 @@ impl Config {
         self.values.metrics.hostname_tag.as_deref()
     }
 
-    /// Returns the global sample rate for all metrics.
-    pub fn metrics_sample_rate(&self) -> f64 {
-        self.values.metrics.sample_rate
-    }
-
-    /// Returns whether local metric aggregation should be enabled.
-    pub fn metrics_aggregate(&self) -> bool {
-        self.values.metrics.aggregate
-    }
-
-    /// Returns whether high cardinality tags should be removed before sending metrics.
-    pub fn metrics_allow_high_cardinality_tags(&self) -> bool {
-        self.values.metrics.allow_high_cardinality_tags
-    }
-
     /// Returns the interval for periodic metrics emitted from Relay.
     ///
     /// `None` if periodic metrics are disabled.
@@ -2597,9 +2586,14 @@ impl Config {
         &self.values.processing.topics.unused
     }
 
-    /// Configuration of the attachment upload service.
-    pub fn upload(&self) -> &UploadServiceConfig {
-        &self.values.processing.upload
+    /// Configuration of the objectstore service.
+    pub fn objectstore(&self) -> &ObjectstoreServiceConfig {
+        &self.values.processing.objectstore
+    }
+
+    /// Configuration of the upload service.
+    pub fn upload(&self) -> &Upload {
+        &self.values.upload
     }
 
     /// Redis servers to connect to for project configs, cardinality limits,