+++ relay-server/src/services/buffer/envelope_buffer/mod.rs
@@ -79,13 +79,13 @@ impl PolymorphicEnvelopeBuffer {
     /// Adds an envelope to the buffer.
     pub async fn push(&mut self, envelope: Box<Envelope>) -> Result<(), EnvelopeBufferError> {
         relay_statsd::metric!(
-            distribution(RelayDistributions::BufferEnvelopeBodySize, sample = 0.01) =
+            distribution(RelayDistributions::BufferEnvelopeBodySize) =
                 envelope.items().map(Item::len).sum::<usize>() as u64,
             partition_id = self.partition_tag()
         );
 
         relay_statsd::metric!(
-            timer(RelayTimers::BufferPush, sample = 0.01),
+            timer(RelayTimers::BufferPush),
             partition_id = self.partition_tag(),
             {
                 match self {
@@ -99,7 +99,7 @@ impl PolymorphicEnvelopeBuffer {
     /// Returns a reference to the next-in-line envelope.
     pub async fn peek(&mut self) -> Result<Peek, EnvelopeBufferError> {
         relay_statsd::metric!(
-            timer(RelayTimers::BufferPeek, sample = 0.01),
+            timer(RelayTimers::BufferPeek),
             partition_id = self.partition_tag(),
             {
                 match self {
@@ -113,7 +113,7 @@ impl PolymorphicEnvelopeBuffer {
     /// Pops the next-in-line envelope.
     pub async fn pop(&mut self) -> Result<Option<Box<Envelope>>, EnvelopeBufferError> {
         relay_statsd::metric!(
-            timer(RelayTimers::BufferPop, sample = 0.01),
+            timer(RelayTimers::BufferPop),
             partition_id = self.partition_tag(),
             {
                 match self {